backend/requirements.txt
Django==4.2.16
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
channels==4.1.0
channels-redis==4.2.0
daphne==4.1.2
django-cors-headers==4.3.1
django-axes==6.5.0
django-environ==0.11.2
django-filter==24.2
drf-spectacular==0.27.2
psycopg2-binary==2.9.9
redis==5.0.8
celery==5.3.6
django-celery-beat==2.6.0
cryptography==42.0.8
Pillow==10.4.0
gunicorn==21.2.0
uvicorn==0.30.6
websockets==13.1
whitenoise==6.7.0
django-db-geventpool==4.0.1
requests==2.32.3
pydantic==2.11.3
bcrypt==4.2.0
django-storages==1.14.4
boto3==1.35.42
